django/cruzroja/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.122
botocore==1.29.122
certifi==2022.12.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
joblib==1.2.0
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.0.33
nltk==3.8.1
numpy==1.24.3
openai==0.27.5
pandas==2.0.1
protobuf==3.20.3
psycopg2-binary==2.9.6
py4j==0.10.9.7
PyJWT==2.6.0
pyspark==3.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
regex==2023.5.5
requests==2.29.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
